feat: Add merino-common package and move the version util to it"
