+++ merino-common/merino_common/__init__.py
@@ -1,0 +1,1 @@
+"""The home for Merino common modules"""
