+++ merino-common/pyproject.toml
@@ -1,0 +1,17 @@
+[project]
+name = "merino-common"
+version = "0.1.0"
+description = "Common Modules for Merino"
+authors = [{ name = "Mozilla" }]
+requires-python = "~=3.14.0"
+dependencies = [
+    "pydantic>=2.13.0,<3",
+]
+
+[tool.uv.build-backend]
+module-name = "merino_common"
+module-root = ""
+
+[build-system]
+requires = ["uv_build>=0.11.6,<0.12.0"]
+build-backend = "uv_build"
