+++ merino/web/dockerflow.py
@@ -5,7 +5,7 @@ import logging
 from fastapi import APIRouter, HTTPException
 from fastapi.responses import RedirectResponse, Response
 
-from merino.utils.version import Version, fetch_app_version_from_file
+from merino_common.utils.version import Version, fetch_app_version_from_file
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
